Preparing to move out render functions to common package
