+++ packages/web-maxrohde-com/src/state/staticFiles.json
@@ -3,7 +3,7 @@
     "names": [
       "tailwind.css"
     ],
-    "generatedName": "tailwind.94d1dd7c2956f6c355ecd76f5d70f900.css"
+    "generatedName": "tailwind.94aef258e1f56fd10882cedef97dc2d9.css"
   },
   {
     "names": [
@@ -98,14 +98,14 @@
       "maxrohde-dev--__index.map",
       "maxrohde-prod--__index.map"
     ],
-    "generatedName": "-__index.cd5cfa91b5c1f0dadfbc328c48f94583.map.json"
+    "generatedName": "-__index.d6a114610733886e6a3c83cc15560717.map.json"
   },
   {
     "names": [
       "maxrohde-dev--__index-bundle.js",
       "maxrohde-prod--__index-bundle.js"
     ],
-    "generatedName": "-__index-bundle.ad9a5d9e59cd17365e5345a31f9d35c8.js"
+    "generatedName": "-__index-bundle.f3f7ebdf06ccaaec3358578da1383474.js"
   },
   {
     "names": [
@@ -117,21 +117,21 @@
     "names": [
       "maxrohde-dev-about-bundle.js"
     ],
-    "generatedName": "about-bundle.e420073487ee084d09f828aec2a79489.js"
+    "generatedName": "about-bundle.0806969aa33746260dca5f48ac462a36.js"
   },
   {
     "names": [
       "maxrohde-dev-categories-_id_.map",
       "maxrohde-prod-categories-_id_.map"
     ],
-    "generatedName": "categories-_id_.2d7d19f9491a6899ddd4409d5a263a8e.map.json"
+    "generatedName": "categories-_id_.47863016e62028a26a7c430347e1581d.map.json"
   },
   {
     "names": [
       "maxrohde-dev-categories-_id_-bundle.js",
       "maxrohde-prod-categories-_id_-bundle.js"
     ],
-    "generatedName": "categories-_id_-bundle.778d4a80b3bf0b99761dda59bfdc17c3.js"
+    "generatedName": "categories-_id_-bundle.682738932c62ef4777171aa77d81d90b.js"
   },
   {
     "names": [
@@ -152,28 +152,28 @@
       "maxrohde-dev-tags-_id_.map",
       "maxrohde-prod-tags-_id_.map"
     ],
-    "generatedName": "tags-_id_.1402409ea705a5a7ecd0d2aabfe36ae3.map.json"
+    "generatedName": "tags-_id_.cea336c23d1d5dc19b2695f3d7ff4f0c.map.json"
   },
   {
     "names": [
       "maxrohde-dev-tags-_id_-bundle.js",
       "maxrohde-prod-tags-_id_-bundle.js"
     ],
-    "generatedName": "tags-_id_-bundle.5ae5f5fc4cbc29ca14c107f50a7c8fe4.js"
+    "generatedName": "tags-_id_-bundle.282d838bc26d94dbc13d84c6e44cc955.js"
   },
   {
     "names": [
       "maxrohde-dev-_post__.map",
       "maxrohde-prod-_post__.map"
     ],
-    "generatedName": "_post__.816c28db4068fa5703f04083d4874d84.map.json"
+    "generatedName": "_post__.7f61a413f432fa689e5e9aad92b78537.map.json"
   },
   {
     "names": [
       "maxrohde-dev-_post__-bundle.js",
       "maxrohde-prod-_post__-bundle.js"
     ],
-    "generatedName": "_post__-bundle.99395690f58758299538e2949b8be36d.js"
+    "generatedName": "_post__-bundle.15b017c63d87fc79421854770c53982a.js"
   },
   {
     "names": [
